[cleanup] Separate test_utils for nightshade runtime
